When BYE received from server, begin disconnecting immediately
Otherwise, it's possible for the server to send a BYE and drop the
connection ungracefully (or the FCLOSE is never received) and leave
the ClientSession active; Geary will think it's still good and attempt
to perform operations on it.  This change will force the connection to
drop and Geary will reestablish.